+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Methods for adding multiple rules and user-functions to a ruleset at once
+
+### Changed
+
+- Add name method to `UserFunction` trait instead of passing the name into the builder
+
 ## [0.3.0] - 2022-11-29
 
 ### Added